change password logic put under a component to use
